chore: (Exercicios) 03, 04 -  Adiciona enum com messes do ano e estacoes
