+++ T19C-Back-End/bloco-26-TypeScript/dia-01-Intoducao-ao-TypeScript/exercices/package.json
@@ -12,6 +12,8 @@
   "devDependencies": {
     "@tsconfig/node16": "^1.0.3",
     "@types/node": "^18.6.2",
+    "@types/readline-sync": "^1.4.4",
+    "readline-sync": "^1.4.10",
     "typescript": "^4.7.4"
   }
 }
