+++ T19C-Back-End/bloco-26-TypeScript/dia-01-Intoducao-ao-TypeScript/exercices/src/Character.ts
@@ -1,0 +1,39 @@
+type Character = {
+  nickname: string,
+  class: string,
+  stats: { agi:number, str:number, int:number, hp:number, mp:number},
+  createdAt: Date,
+};
+
+const characters: Character[] = [
+  {
+    nickname: 'xKillerx',
+    class: 'warrior',
+    stats: { agi: 50, str: 100, int: 25, hp: 1000, mp: 300 },
+    createdAt: new Date('2003-10-1')
+  },
+  {
+    nickname: 'jainaProud',
+    class: 'mage',
+    stats: { agi: 80, str: 40, int: 150, hp: 630, mp: 1100 },
+    createdAt: new Date('2003-10-2')
+  },
+  {
+    nickname: 'catIn',
+    class: 'hunter',
+    stats: { agi: 150, str: 90, int: 80, hp: 800, mp: 600 },
+    createdAt: new Date('2003-10-15')
+  },
+]
+
+function printCharacter(character: Character, index: number) {
+  const { nickname, class: cls, createdAt, stats } = character;
+  
+  console.log(`\n\n===== Character: ${index + 1} ========`);
+  console.log(`nickname: ${nickname}
+    class: ${cls}
+    createdAt: ${createdAt}
+    stats ${Object.entries(stats)}`);
+}
+
+characters.forEach(printCharacter);
