+++ T19C-Back-End/bloco-26-TypeScript/dia-01-Intoducao-ao-TypeScript/exercices/src/Months.ts
@@ -1,0 +1,16 @@
+enum Months {
+  JANEIRO = "Janeiro",
+  FEVEREIRO = "Fevereiro",
+  MARCO = "Março",
+  ABRIL = "Abril",
+  MAIO = "Maio",
+  JUNHO = "Junho",
+  JULHO = "Julho",
+  AGOSTO = "Agosto",
+  SETEMBRO = "Setembro",
+  OUTUBRO = "Outubro",
+  NOVEMBRO = "Novembro",
+  DEZEMBRO = "Dezembro",
+}
+
+export default Months
