+++ T19C-Back-End/bloco-26-TypeScript/dia-01-Intoducao-ao-TypeScript/exercices/src/Seasons.ts
@@ -1,0 +1,8 @@
+enum Seasons {
+  OUTONO = "Outono",
+  INVERNO = "Inverno",
+  PRIMAVERA = "Primavera",
+  VERAO = "Verão",
+}
+
+export default Seasons
